+++ backend/gonzales/static/assets/trending-up-UDCt6Pfh.js
@@ -1,4 +1,4 @@
-import{c as e}from"./index-BcReyCNX.js";/**
+import{c as e}from"./index-CRbZuQtx.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.